Move getItems helper functions to utils.
